backend/requirements.txt
fastapi>=0.104.0
psutil>=5.9.0
uvicorn[standard]>=0.24.0
nd2>=0.7.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
Pillow>=10.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
